create a new component instance for each component
__Problem:__
When parsing files with multiple components the very same
component instance was used for each component in the file.
That caused _every_ component to be identical to the last component
in the file.

__Solution:__
Instead of using the same component instance for each file_component
a new instance is getting created and added to the `components`-Array