Cache IceCube schedules to improve standard/daylight performance
If a calendar had a standard or daylight start date very far in the
past, enumerating occurrences for those was very slow. This calculation
was not reused for multiple events with the same tzid.

Add a caching mechanism to Daylight/Standard so that occurrences are
only calculated once per calendar.

The cache only calculates the occurrences it needs on demand.